add default group screen list + improve table

+++ src/screen/list-group-item.tsx
@@ -4,14 +4,16 @@ import classNames from "classnames"
 
 import {ReactComponent as IconSettings} from "./icon-settings.svg"
 import {ReactComponent as IconTrash} from "./icon-trash.svg"
+import {ReactComponent as IconArrowDown} from "./icon-arrow-down.svg"
+import {ReactComponent as IconFolder} from "./icon-folder.svg"
 import {Group} from "./model"
 import useScreens from "./context"
 
 import classes from "./list-group-item.module.scss"
 
 type GroupListItemProps = {
-  group: Group
-  onDrop: (group: Group) => Promise<void>
+  group: Group | null
+  onDrop: (group: Group | null) => Promise<void>
 }
 
 const GroupListItem: FC<GroupListItemProps> = props => {
@@ -24,7 +26,7 @@ const GroupListItem: FC<GroupListItemProps> = props => {
   async function deleteGroup(evt: React.MouseEvent<HTMLButtonElement>) {
     evt.stopPropagation()
 
-    if (window.confirm(t("screen:confirm-deletion"))) {
+    if (group && window.confirm(t("screen:confirm-deletion"))) {
       await $screen.deleteGroup(group.id)
     }
   }
@@ -47,40 +49,41 @@ const GroupListItem: FC<GroupListItemProps> = props => {
     await props.onDrop(group)
   }
 
+  const filteredScreens = screens.filter(s => (group ? s.groupId === group.id : s.groupId === null))
+
   return (
     <>
       <tr
-        key={group.id}
-        className={classNames(classes.row, {[classes.draggedOver]: draggedOver})}
+        key={(group && group.id) || "default"}
+        className={classNames(classes.row, {
+          [classes.defaultRow]: !group,
+          [classes.draggedOver]: draggedOver,
+        })}
         onClick={handleClick}
         onDragOver={handleDragOver}
         onDragLeave={handleDragLeave}
         onDrop={handleDrop}
       >
         <td className={classes.selectCol}>
-          <img
-            className={classNames({[classes.open]: open})}
-            src="/images/icon-arrow-down.svg"
-            alt=""
-          />
+          <IconArrowDown className={classNames({[classes.open]: open})} />
         </td>
         <td className={classes.linkCol}>
-          ({t("screen:n-screens", {n: screens.filter(s => s.groupId === group.id).length})})
+          ({t("screen:n-screens", {n: filteredScreens.length}).toLowerCase()})
         </td>
-        <td className={classes.statusCol}>
-          <img src="/images/icon-folder.svg" alt="" />
-        </td>
-        <td className={classes.nameCol}>{group.name}</td>
+        <td className={classes.statusCol}>{group && <IconFolder />}</td>
+        <td className={classes.nameCol}>{group && group.name}</td>
         <td className={classes.layoutCol} />
         <td className={classes.actionsCol}>
-          <div>
-            <button type="button">
-              <IconSettings className={classes.icon} />
-            </button>
-            <button type="button" onClick={deleteGroup}>
-              <IconTrash className={classes.icon} />
-            </button>
-          </div>
+          {group && (
+            <div>
+              <button type="button">
+                <IconSettings className={classes.icon} />
+              </button>
+              <button type="button" onClick={deleteGroup}>
+                <IconTrash className={classes.icon} />
+              </button>
+            </div>
+          )}
         </td>
       </tr>
       {open && children}
